Crate wasmi_arena

Fast arena allocators for different usage purposes.

They cannot deallocate single allocated entities for extra efficiency. These allocators mainly serve as the backbone for an efficient Wasm store implementation.

Structs

Arena

An arena allocator with a given index and entity type.

ComponentVec

Stores components for entities backed by a Vec.

DedupArena

A deduplicating arena allocator with a given index and entity type.

GuardedEntity

A guarded entity.

Iter

An iterator over shared references of arena entities and their indices.

IterMut

An iterator over exclusive references of arena entities and their indices.

Traits

ArenaIndex

Types that can be used as indices for arenas.